Export json lists to data folder
- Create data folder for JSON
- Move lists into data folder
- Combine projectCards and project lists
