+++ src/components/index/projects/FeaturedProjects.js
@@ -1,34 +1,8 @@
 import { useMediaQuery } from "react-responsive";
 import { motion } from "framer-motion";
 
+import cardsData from "../../../data/cardsData";
 import { GithubLink, DeployLink } from "./SocialLinks";
-
-const cardsData = [
-  {
-    image: "portfolio-thumbnail.webp",
-    name: "portfolio",
-    tagline: "Portfolio Website",
-    heading: "Portfolio Website",
-    description:
-      "My portfolio website which includes my latest projects and serves as a point of contact for potential clients and employers, it's constantly updated to reflect my development progress.",
-    technologies: ["Handlebars/React", "ExpressJS", "NodeJS", "CSS/Tailwind"],
-    githubLink: "https://github.com/JoshHaywood/portfolio-website",
-    deployLink: "https://www.joshhaywood-portfolio.com/",
-    alignment: "right",
-  },
-  {
-    image: "tech-terminus-thumbnail.jpg",
-    name: "tech-terminus",
-    tagline: "Ecommerce Website",
-    heading: "Dissertation project",
-    description:
-      "Artefact A of an A/B test used as part of my scientific study on the impact of interfaces and design on user experiences in web-based applications.",
-    technologies: ["React", "ExpressJS", "NodeJS", "Axios", "MySQL"],
-    githubLink: "https://github.com/JoshHaywood/tech-terminus",
-    deployLink: "http://www.tech-terminus.me/",
-    alignment: "left",
-  },
-];
 
 export default function FeaturedProjects(props) {
   const { sidebar, setSidebar, setProjectName } = props;
