+++ src/data/components.js
@@ -1,0 +1,13 @@
+import Skills from "../components/index/Skills";
+import About from "../components/index/About";
+import Projects from "../components/index/projects/Projects";
+import Contact from "../components/index/Contact";
+
+const components = [
+  { name: "About", component: About },
+  { name: "Skills", component: Skills },
+  { name: "Projects", component: Projects },
+  { name: "Contact", component: Contact },
+];
+
+export default components;
