+++ src/data/links.js
@@ -1,0 +1,8 @@
+const links = [
+  { id: "about", name: "About" },
+  { id: "skills", name: "Skills" },
+  { id: "projects", name: "Projects" },
+  { id: "contact", name: "Contact" },
+];
+
+export default links;
