+++ src/data/projects.js
@@ -1,0 +1,207 @@
+const projects = [
+  {
+    projectName: "portfolio",
+    repoLink: "https://github.com/JoshHaywood/portfolio-website",
+    deployLink: "https://www.joshhaywood-portfolio.com/",
+    heading: "Personal Portfolio Website",
+    previewText:
+      "A website showcasing a curated selection of my projects and providing easy access to my contact information for potential employers.",
+    previewStructure: [
+      {
+        item: "React",
+      },
+      {
+        item: "TailwindCSS",
+      },
+      {
+        item: "Framer Motion",
+      },
+    ],
+    tagline: "Highlighting my skills and projects",
+    projectImage: "../images/portfolio-thumbnail.webp",
+    overview:
+      "This project was a website that I created to use as a marketing tool to showcase my relevant experience and various projects that I have undertaken. The goal was to demonstrate my range of abilities to potential employers or people of interest, and to serve as a hub for my various social profiles on Github, LinkedIn, and email.",
+    structure: [
+      "HTML",
+      "Handlebars",
+      "CSS",
+      "Javascript",
+      "Express.js",
+      "Node.js",
+    ],
+    role: `I created a personal portfolio website to showcase my skills and experience to potential employers. I faced challenges with Handlebars and difficulties with the design of the page layouts, but overcame them by converting to React, researching other developers' work and using Framer Motion for animations. The end result was a sleek and visually appealing website that effectively highlighted my abilities and experience.`,
+  },
+  {
+    projectName: "tech-terminus",
+    repoLink: "https://github.com/JoshHaywood/tech-terminus",
+    deployLink: "http://www.tech-terminus.me/",
+    heading: "Ecommerce website",
+    previewText:
+      "An Ecommerce website that I developed during my research and development dissertation, showcasing standard ecommerce features using a restful API.",
+    previewStructure: [
+      {
+        item: "React",
+      },
+      {
+        item: "Express",
+      },
+      {
+        item: "MySQL",
+      },
+    ],
+    tagline: "Exploring the science of UX through ecommerce Design",
+    projectImage: "../images/tech-terminus-index.png",
+    overview:
+      "This was an ecommerce website that served as an artefact in my research into how less experienced developers could use design to improve the user experience of their applications with limited knowledge of developing.",
+    structure: [
+      "HTML",
+      "React",
+      "Tailwinds CSS",
+      "Material UI",
+      "Javascript",
+      "Express.js",
+      "Node.js",
+      "MySQL",
+    ],
+    role: "I designed, developed, and hosted the application from the ground up. This involved building all the core features, creating the front-end, writing endpoints for the back-end, creating database tables and hosting the site with Heroku. Additionally, as part of my study, I conducted an A/B test comparing this site to one made with a website builder. After that, I recruited participants for a qualitative study and presented my findings to a panel of academics.",
+  },
+  {
+    projectName: "realtime-messaging-app",
+    repoLink: "https://github.com/JoshHaywood/messaging-app",
+    deployLink: "https://messaging-app.herokuapp.com/",
+    heading: "Real-Time Messaging App",
+    previewText:
+      "An upgraded version of my previous real-time messaging app, showcasing improvements and new features. The app allows users to exchange messages instantly and interact in real-time.",
+    previewStructure: [
+      {
+        item: "Next.js",
+      },
+      {
+        item: "Express",
+      },
+      {
+        item: "TypeScript",
+      },
+    ],
+    tagline: "Real-time communication, at your fingertips",
+    projectImage: "../images/messaging-app-thumbnail.png",
+    overview:
+      "This project is the second iteration of my real-time messaging app, aimed at improving the user experience and adding new features based on user feedback from the initial version. The app allows users to communicate in real-time, providing a seamless and efficient messaging experience.",
+    structure: [
+      "HTML",
+      "Next.js",
+      "Tailwinds CSS",
+      "Material UI",
+      "TypeScript",
+      "Express.js",
+      "Node.js",
+      "Socket.io",
+      "MySQL",
+    ],
+    role: `As the sole developer of this project, I took on the challenge of enhancing the app's functionality and design. I implemented new features such as friend requests, profiles and settings, and time stamping messages, improving the overall user experience. I also optimized the codebase for better performance and scalability with Next and TypeScript. Throughout the development process, I actively sought user feedback and iterated on the design to meet their needs and preferences.`,
+  },
+  {
+    projectName: "freelancing-website",
+    repoLink: "https://github.com/JoshHaywood/pop-up-team",
+    deployLink: "/error",
+    heading: "Freelancing Website",
+    previewText:
+      "A collaborative project I worked on as part of a group, where we developed a freelancing website to connect clients and freelancers.",
+    previewStructure: [
+      {
+        item: "React",
+      },
+      {
+        item: "Express",
+      },
+      {
+        item: "mySQL",
+      },
+    ],
+    tagline: "Connecting clients and freelancers",
+    projectImage: "../images/popup-team-thumbnail.png",
+    overview:
+      "This project was a freelancing website that I developed as part of a group project. The goal was to create a platform that connects clients with freelancers across various technology disciplines. The website provided features such as a client directory, instant messaging, and a built-in kanban board.",
+    structure: [
+      "HTML",
+      "React",
+      "Tailwinds CSS",
+      "Material UI",
+      "Javascript",
+      "Express.js",
+      "Node.js",
+      "Socket.io",
+      "MySQL",
+    ],
+    role: "In this project, I played a key role in full-stack development. I was responsible for creating interfaces, database tables, and endpoints for core features. This involved developing user registration and authentication systems, designing the client directory and freelancer profile pages, and integrating a real-time one-to-one instant messaging system. The final result was a user-friendly platform that effectively facilitated the connection between clients and freelancers.",
+  },
+  {
+    projectName: "student-survival-store",
+    repoLink: "https://github.com/JoshHaywood/student-survival-store",
+    deployLink: "/error",
+    heading: "SaaS Student Survival Store",
+    previewText:
+      "A group project that involved creating a software as a service (SaaS) product utilizing a monthly subscription model.",
+    previewStructure: [
+      {
+        item: "React",
+      },
+      {
+        item: "Express",
+      },
+      {
+        item: "Sqlite3",
+      },
+    ],
+    tagline: "Surviving student life, made easy",
+    projectImage: "../images/student-survival-store-index.webp",
+    overview:
+      "This was a group project in my second year of university that centered around a business that provided a subscription-based box of student-related supplies. My role was front-end development and documentation handling. I maintained our team's agile approach by documenting each stand-up's notes, targets, target status, and any issues that occurred for that week. This project was my first time using React, so my development was not as substantial as normal, but I did create components such as the mobile header and implemented the user profile modal.",
+    structure: [
+      "HTML",
+      "React",
+      "Tailwinds CSS",
+      "Javascript",
+      "Express.js",
+      "Node.js",
+      "SQLite",
+    ],
+    role: "I helped create an e-commerce website to sell student survival kits. My role was focused on the development and design of the website, as well as working with a team to create the product offerings. One of the main challenges was ensuring a smooth and user-friendly checkout process, which was addressed by implementing a variety of payment options and testing them thoroughly.",
+  },
+  {
+    projectName: "esports-center",
+    repoLink: "https://github.com/JoshHaywood/esports-center",
+    deployLink: "/error",
+    heading: "Esports Stats Tracking Website",
+    previewText:
+      "A statistics tracking tool for Esports, which I led as the project manager, displaying my leadership and interest in the rising Esports industry.",
+    previewStructure: [
+      {
+        item: "React",
+      },
+      {
+        item: "Express",
+      },
+      {
+        item: "MySQL",
+      },
+    ],
+    tagline: "Leading the way in Esports statistics",
+    projectImage: "../images/tech-terminus-thumbnail.jpg",
+    overview:
+      "This was a group project in my second year with the same team as the Student Survival Store project. Similarly to that project, the Esports Center was first presented through a business plan in which each team member pitched their own ideas. My own pitch was for the Esports Center, a statistics tracking website focused on the sporting genre of esports, or the competitive play of video games. My role in the project was front-end development and design, as well as documentation.",
+    structure: [
+      "HTML",
+      "React",
+      "CSS",
+      "Material UI",
+      "Javascript",
+      "Express.js",
+      "Node.js",
+      "MySQL",
+    ],
+    role: `I assisted in creating a web platform for a local e-sport center, where users can find information about tournaments and events, as well as register and pay for them online. My role included working on the design and development of the website, as well as implementing a system for tournament registration and payment. One of the main challenges was ensuring the security and reliability of the payment system, which was addressed by using a well-known and secure payment gateway.`,
+  },
+];
+
+export default projects;
