+++ src/data/technologies.js
@@ -1,0 +1,21 @@
+const technologies = [
+  {
+    listColumn: [{ label: "React" }, { label: "NextJS" }, { label: "SASS" }],
+  },
+  {
+    listColumn: [
+      { label: "TailwindCSS" },
+      { label: "TypeScript" },
+      { label: "Node.js" },
+    ],
+  },
+  {
+    listColumn: [
+      { label: "Web Sockets" },
+      { label: "Express" },
+      { label: "Heroku" },
+    ],
+  },
+];
+
+export default technologies;
