+++ src/pages/Index.js
@@ -1,17 +1,7 @@
 import { Helmet } from "react-helmet";
 
+import components from "../data/components.js";
 import Hero from "../components/index/hero/Hero";
-import Skills from "../components/index/Skills";
-import About from "../components/index/About";
-import Projects from "../components/index/projects/Projects";
-import Contact from "../components/index/Contact";
-
-const components = [
-  { name: "About", component: About },
-  { name: "Skills", component: Skills },
-  { name: "Projects", component: Projects },
-  { name: "Contact", component: Contact },
-];
 
 export default function Index(props) {
   const { sidebar, setSidebar, projectName, setProjectName } = props;
